Bumping up net framework

+++ Container/Extension/Extension.v4.cs
@@ -69,7 +69,7 @@ namespace Container
         public void ExtensionCanAddStrategy()
         {
             SpyStrategy spy = new SpyStrategy();
-            SpyExtension extension = new SpyExtension(spy, UnityBuildStage.PostInitialization);
+            SpyExtension extension = new SpyExtension(spy, (int)UnityBuildStage.PostInitialization);
 
             IUnityContainer container = new UnityContainer()
                 .AddExtension(extension);
@@ -86,7 +86,7 @@ namespace Container
             SpyPolicy spyPolicy = new SpyPolicy();
 
             SpyExtension extension =
-                new SpyExtension(spy, UnityBuildStage.PostInitialization, spyPolicy, typeof(SpyPolicy));
+                new SpyExtension(spy, (int)UnityBuildStage.PostInitialization, spyPolicy, typeof(SpyPolicy));
 
             IUnityContainer container = new UnityContainer()
                 .AddExtension(extension);
